* develop: (68 commits)
  Update CHANGELOG [ci skip]
  Fix of wrong include symbol of CCVector.h
  closed #3290: [JSB] The order of onEnter and onExit is wrong. For develop branch.
  Update CHANGELOG [ci skip]
  Update AUTHORS [ci skip]
  Added test case for binary fnt file
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Small fix for `FREE` wrong variable.
  More warning fixes.
  AngelCode binary file format support for LabelBMFont
  [AUTO] : updating submodule reference to latest autogenerated bindings
  Bug fixes for cocostudio.
  Update AUTHORS [ci skip]
  Update CHANGELOG [ci skip]
  Improve the coding style
  Modify wrong codes
  [AUTO] : updating submodule reference to latest autogenerated bindings
  fix failed to receive dispatch signal on some android device,such as [HTC EVO 3D X515m].
  Fixed bugs
  [AUTO] : updating submodule reference to latest autogenerated bindings
  ...

+++ cocos/2d/CCAtlasNode.h
@@ -63,8 +63,8 @@ public:
     void setTextureAtlas(TextureAtlas* textureAtlas);
     TextureAtlas* getTextureAtlas() const;
     
-    void setQuadsToDraw(int quadsToDraw);
-    int getQuadsToDraw() const;
+    void setQuadsToDraw(ssize_t quadsToDraw);
+    ssize_t getQuadsToDraw() const;
 
     
     // Overrides
@@ -126,7 +126,7 @@ protected:
     BlendFunc _blendFunc;
 
     // quads to draw
-    int _quadsToDraw;
+    ssize_t _quadsToDraw;
     // color uniform
     GLint    _uniformColor;
     // This varible is only used for LabelAtlas FPS display. So plz don't modify its value.
